requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
google-generativeai==0.7.2
